feat: add example scripts for Spot and Futures API interactions via REST and WebSocket clients
