+++ examples/Auth/fasterHmacSign.ts
@@ -1,4 +1,6 @@
-import { RestClient } from '../src';
+import { createHmac } from 'crypto';
+
+import { RestClient } from '../../src/index.js';
 // import from npm, after installing via npm `npm install bitmart-api`
 // import { RestClient } from 'bitmart-api';
 
@@ -12,6 +14,13 @@ const client = new RestClient({
   apiKey: account.key,
   apiSecret: account.secret,
   apiMemo: account.memo,
+  /**
+   * Overkill in almost every case, but if you need any optimisation available,
+   * you can inject a faster sign mechanism such as node's native createHmac:
+   */
+  customSignMessageFn: async (message, secret) => {
+    return createHmac('sha256', secret).update(message).digest('hex');
+  },
 });
 
 async function getSpotBalances() {
@@ -20,7 +29,7 @@ async function getSpotBalances() {
 
     console.log('Balances: ', JSON.stringify(balances, null, 2));
   } catch (e) {
-    console.error(`Req error: `, e);
+    console.error('Req error: ', e);
   }
 }
