+++ examples/Rest/Futures/futures-get-balances.ts
@@ -1,4 +1,4 @@
-import { FuturesClientV2 } from '../src/index.js';
+import { FuturesClientV2 } from '../../../src/index.js';
 // // import from npm, after installing via npm `npm install bitmart-api`
 // import { FuturesClientV2 } from 'bitmart-api';
 
@@ -20,7 +20,7 @@ async function getFuturesAssets() {
 
     console.log('Balances: ', JSON.stringify(balances, null, 2));
   } catch (e) {
-    console.error(`Req error: `, e);
+    console.error('Req error: ', e);
   }
 }
