+++ examples/Rest/Futures/futures-get-klines.ts
@@ -1,4 +1,4 @@
-import { FuturesClientV2 } from '../src/index.js';
+import { FuturesClientV2 } from '../../../src/index.js';
 // // import from npm, after installing via npm `npm install bitmart-api`
 // import { FuturesClientV2 } from 'bitmart-api';
 
@@ -14,7 +14,7 @@ async function getFuturesKlines() {
 
     console.log('Tickers: ', JSON.stringify(klines, null, 2));
   } catch (e) {
-    console.error(`Req error: `, e);
+    console.error('Req error: ', e);
   }
 }
