+++ examples/Rest/Futures/futures-get-tickers.ts
@@ -1,4 +1,4 @@
-import { FuturesClientV2 } from '../src/index.js';
+import { FuturesClientV2 } from '../../../src/index.js';
 // // import from npm, after installing via npm `npm install bitmart-api`
 // import { FuturesClientV2 } from 'bitmart-api';
 
@@ -10,7 +10,7 @@ async function getFuturesTickers() {
 
     console.log('Tickers: ', JSON.stringify(tickers, null, 2));
   } catch (e) {
-    console.error(`Req error: `, e);
+    console.error('Req error: ', e);
   }
 }
