+++ examples/Websocket/ws-spot-public.ts
@@ -1,7 +1,4 @@
-import {
-  WebsocketClient,
-  // WsSpotOperation,
-} from '../src';
+import { WebsocketClient } from '../../src/index.js';
 
 // import from npm, after installing via npm `npm install bitmart-api`
 // import { WebsocketClient } from 'bitmart-api';
@@ -92,7 +89,7 @@ async function start() {
       'spot',
     );
   } catch (e) {
-    console.error(`Req error: `, e);
+    console.error('Req error: ', e);
   }
 }
